Extract `CoreData` managed models in their own files
